去除无用文件

+++ src/main/java/demo/ms/todo/repository/CardRepository.java
@@ -3,5 +3,5 @@ package demo.ms.todo.repository;
 import demo.ms.common.entity.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card,Integer> {
+public interface CardRepository extends JpaRepository<Card,Long> {
 }
